Refactor state types and initial functions into one file
