+++ src/components/MultiTagCell/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Tag as TagState } from "src/services/appData/state/tag";
+import { Tag as TagState } from "src/services/appData/state/types";
 import Tag from "../Tag";
 interface Props {
 	tags: TagState[];
