--- src/services/appData/state/row.ts
@@ -1,17 +1,0 @@
-export interface Row {
-	id: string;
-	initialIndex: number;
-	creationTime: number;
-}
-
-export const initialRow = (
-	id: string,
-	initialIndex: number,
-	creationTime: number
-): Row => {
-	return {
-		id,
-		initialIndex,
-		creationTime,
-	};
-};
